Revise the structure to make them clean.
