+++ app/src/main/java/com/dragonnetwork/hihealth/medication/MedicationAdaptorOLD.java
@@ -1,0 +1,25 @@
+package com.dragonnetwork.hihealth.medication;
+
+
+import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
+
+public class MedicationAdaptorOLD extends RecyclerView.Adapter {
+    @NonNull
+    @Override
+    public RecyclerView.ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
+        return null;
+    }
+
+    @Override
+    public void onBindViewHolder(@NonNull RecyclerView.ViewHolder holder, int position) {
+
+    }
+
+    @Override
+    public int getItemCount() {
+        return 0;
+    }
+}
